+++ changelog.md
@@ -6,6 +6,19 @@ Version format based on <http://semver.org/>
 
 ## [Unreleased]
 
+## [2.1.0.0] - 2020-12-09
+
+### Changed
+
+- Deprecated method im_open replaced by conversations_open
+- Deprecated method groups_info replaced by conversations_info
+- Deprecated method channels_info replaced by conversations_info
+- Deprecated API method channels.list replaced by conversations.list
+- Deprecated API method groups.list replaced by conversations.list
+- Deprecated API method mpim.list replaced by conversations.list
+- Deprecated API method im.list replaced by conversations.list
+- Deprecated API method channels.setTopic replaced by conversations.setTopic
+
 ## [2.0.0.0] - 2020-10-29
 
 Using v 2 to indicate this is no longer the https://github.com/litaio/lita-slack version
@@ -17,7 +30,7 @@ Using v 2 to indicate this is no longer the https://github.com/litaio/lita-slack
 
 ### Changed
 
-- Strip No break spaces in message texts
+- Strip No-break spaces in message texts
 - Add API methods for identifying private chats
   - channel_info: Deprecated API, will need to be removed soon.
   - group_info : Deprecated API, will need to be removed soon.